Spot light

+++ OpenGLCourse/OpenGLCourse/PointLight.cpp
@@ -6,7 +6,6 @@ PointLight::PointLight() : Light()
 	constant = 1.0f;
 	linear = 0.0f;
 	exponent = 0.0f;
-
 }
 
 PointLight::PointLight(GLfloat red, GLfloat green, GLfloat blue, 
@@ -20,9 +19,9 @@ PointLight::PointLight(GLfloat red, GLfloat green, GLfloat blue,
 	exponent = exp;
 }
 
-void PointLight::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColourLocation,
-					GLfloat diffuseIntensityLocation, GLfloat positionLocation,
-					GLfloat constantLocation, GLfloat linearLocation, GLfloat exponentLocation)
+void PointLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation,
+		GLuint diffuseIntensityLocation, GLuint positionLocation,
+		GLuint constantLocation, GLuint linearLocation, GLuint exponentLocation)
 {
 		glUniform3f(ambientColourLocation, colour.x, colour.y, colour.z);
 		glUniform1f(ambientIntensityLocation, ambientIntensity);
